+++ ktrain/text/preprocessor.py
@@ -356,7 +356,7 @@ class TextPreprocessor(Preprocessor):
         self.lang = lang
         self.multilabel = multilabel
 
-
+    
     def get_preprocessor(self):
         raise NotImplementedError
 
@@ -370,7 +370,7 @@ class TextPreprocessor(Preprocessor):
 
 
     def set_multilabel(self, data, mode):
-        if mode == 'train':
+        if mode == 'train' and self.get_classes():
             self.multilabel = U.is_multilabel(data)
 
 
@@ -400,6 +400,56 @@ class TextPreprocessor(Preprocessor):
             seg_list = list(seg_list)
             split_texts.append(seg_list)
         return [" ".join(tokens) for tokens in split_texts] 
+
+
+    @classmethod
+    def seqlen_stats(cls, list_of_texts):
+        """
+        compute sequence length stats from
+        list of texts in any spaces-segmented language
+        Args:
+            list_of_texts: list of strings
+        Returns:
+            dict: dictionary with keys: mean, 95percentile, 99percentile
+        """
+        counts = []
+        for text in list_of_texts:
+            if isinstance(text, (list, np.ndarray)):
+                lst = text
+            else:
+                lst = text.split()
+            counts.append(len(lst))
+        p95 = np.percentile(counts, 95)
+        p99 = np.percentile(counts, 99)
+        avg = sum(counts)/len(counts)
+        return {'mean':avg, '95percentile': p95, '99percentile':p99}
+
+
+    def print_seqlen_stats(self, texts, mode, verbose=1):
+        """
+        prints stats about sequence lengths
+        """
+        if verbose and not self.is_nospace_lang():
+            stat_dict = TextPreprocessor.seqlen_stats(texts)
+            print( "%s sequence lengths:" % mode)
+            for k in stat_dict:
+                print("\t%s : %s" % (k, int(round(stat_dict[k]))))
+
+
+    def _transform_y(self, y_data):
+        """
+        preprocess y
+        If shape of y is 1, then task is considered classification if self.c exists
+        or regression if not.
+        """
+        if y_data is None: return y_data
+        # if shape is 1, this is either a classification or regression task 
+        # depending on class_names existing
+        y_data = np.array(y_data) if type(y_data) == list else y_data
+        y_data = to_categorical(y_data) if len(y_data.shape) == 1 and self.get_classes() else y_data
+        return y_data
+
+
 
 
 
@@ -439,6 +489,8 @@ class StandardTextPreprocessor(TextPreprocessor):
         """
         preprocess training set
         """
+        if self.lang is None: self.lang = detect_lang(train_text)
+
 
         U.vprint('language: %s' % (self.lang), verbose=verbose)
 
@@ -454,19 +506,19 @@ class StandardTextPreprocessor(TextPreprocessor):
         # convert to word IDs
         x_train = self.tok.texts_to_sequences(train_text)
         U.vprint('{} train sequences'.format(len(x_train)), verbose=verbose)
-        U.vprint('Average train sequence length: {}'.format(np.mean(list(map(len, x_train)), dtype=int)), verbose=verbose)
+        self.print_seqlen_stats(x_train, 'train', verbose=verbose)
 
         # add ngrams
         x_train = self._fit_ngrams(x_train, verbose=verbose)
 
         # pad sequences
         x_train = sequence.pad_sequences(x_train, maxlen=self.maxlen)
+        U.vprint('x_train shape: ({},{})'.format(x_train.shape[0], x_train.shape[1]), verbose=verbose)
 
         # transform y
-        if len(y_train.shape) == 1:
-            y_train = to_categorical(y_train)
-        U.vprint('x_train shape: ({},{})'.format(x_train.shape[0], x_train.shape[1]), verbose=verbose)
-        U.vprint('y_train shape: ({},{})'.format(y_train.shape[0], y_train.shape[1]), verbose=verbose)
+        y_train = self._transform_y(y_train)
+        if y_train is not None and verbose:
+            print('y_train shape: %s' % (y_train.shape))
 
         # return
         result =  (x_train, y_train)
@@ -487,8 +539,7 @@ class StandardTextPreprocessor(TextPreprocessor):
         # convert to word IDs
         x_test = self.tok.texts_to_sequences(test_text)
         U.vprint('{} test sequences'.format(len(x_test)), verbose=verbose)
-        U.vprint('Average test sequence length: {}'.format(np.mean(list(map(len, x_test)), 
-                                                                 dtype=int)), verbose=verbose)
+        self.print_seqlen_stats(x_test, 'test', verbose=verbose)
 
         # add n-grams
         x_test = self._add_ngrams(x_test, mode='test', verbose=verbose)
@@ -498,11 +549,11 @@ class StandardTextPreprocessor(TextPreprocessor):
         x_test = sequence.pad_sequences(x_test, maxlen=self.maxlen)
         U.vprint('x_test shape: ({},{})'.format(x_test.shape[0], x_test.shape[1]), verbose=verbose)
 
-        # convert y
-        if y_test is not None:
-            if len(y_test.shape) == 1:
-                y_test = to_categorical(y_test)
-            U.vprint('y_test shape: ({},{})'.format(y_test.shape[0], y_test.shape[1]), verbose=verbose)
+        # transform y
+        y_test = self._transform_y(y_test)
+        if y_test is not None and verbose:
+            print('y_test shape: %s' % (y_test.shape))
+
 
         # return
         return (x_test, y_test)
@@ -555,6 +606,7 @@ class StandardTextPreprocessor(TextPreprocessor):
             new_sequences.append(new_list)
         U.vprint('Average {} sequence length with ngrams: {}'.format(mode,
             np.mean(list(map(len, new_sequences)), dtype=int)), verbose=verbose)    
+        self.print_seqlen_stats(new_sequences, '%s (w/ngrams)' % mode, verbose=verbose)
         return new_sequences
 
 
@@ -626,14 +678,14 @@ class BERTPreprocessor(TextPreprocessor):
         """
         if mode == 'train' and y is None:
             raise ValueError('y is required when mode=train')
+        if self.lang is None and mode=='train': self.lang = detect_lang(texts)
         U.vprint('preprocessing %s...' % (mode), verbose=verbose)
         U.vprint('language: %s' % (self.lang), verbose=verbose)
 
         x = bert_tokenize(texts, self.tok, self.maxlen, verbose=verbose)
-        if y is not None:
-            if len(y.shape) == 1:
-                y = to_categorical(y)
-            #U.vprint('\ty shape: ({},{})'.format(y.shape[0], y.shape[1]), verbose=verbose)
+
+        # transform y
+        y = self._transform_y(y)
         result = (x, y)
         self.set_multilabel(result, mode)
         return result
@@ -700,15 +752,16 @@ class TransformersPreprocessor(TextPreprocessor):
         preprocess training set
         """
         U.vprint('preprocessing %s...' % (mode), verbose=verbose)
+        if self.lang is None and mode=='train': self.lang = detect_lang(texts)
         U.vprint('language: %s' % (self.lang), verbose=verbose)
+        self.print_seqlen_stats(texts, mode, verbose=verbose)
+
+        # transform y
         if y is None and mode == 'train':
             raise ValueError('y is required for training sets')
         elif y is None:
             y = np.array([1] * len(texts))
-        if y is not None:
-            if len(y.shape) == 1:
-                y = to_categorical(y)
-
+        y = self._transform_y(y)
         dataset = hf_convert_examples(texts, y=y, tokenizer=self.tok, max_length=self.maxlen,
                                       pad_on_left=bool(self.name in ['xlnet']),
                                       pad_token=self.tok.convert_tokens_to_ids([self.tok.pad_token][0]),
@@ -775,8 +828,8 @@ class Transformer(TransformersPreprocessor):
         """
         if not use_with_learner and batch_size is None:
             raise ValueError('batch_size is required when use_with_learner=False')
-        if classes is None or not classes:
-            raise ValueError('classes argument is required - provide list of class names as strings')
+        if multilabel and (classes is None or not classes):
+            raise ValueError('classes argument is required when multilabel=True')
         super().__init__(model_name,
                          maxlen, max_features=10000, classes=classes, multilabel=multilabel)
         self.batch_size = batch_size
@@ -827,6 +880,9 @@ class Transformer(TransformersPreprocessor):
 
 
     def get_classifier(self):
+        if not self.get_classes():
+            warnings.warn('no class labels were provided - treating as regression')
+            return self.get_regression_model()
         num_labels = len(self.get_classes())
         model = self.model_type.from_pretrained(self.model_name, num_labels=num_labels)
         if self.multilabel:
@@ -837,6 +893,25 @@ class Transformer(TransformersPreprocessor):
                       optimizer=keras.optimizers.Adam(learning_rate=3e-5, epsilon=1e-08),
                       metrics=['accuracy'])
         return model
+
+
+    def get_regression_model(self):
+        if self.get_classes():
+            warnings.warn('class labels were provided - treating as classification problem')
+            return self.get_classifier()
+        model = self.model_type.from_pretrained(self.model_name, num_labels=1)
+        loss_fn = 'mse'
+        model.compile(loss=loss_fn,
+                      optimizer=keras.optimizers.Adam(learning_rate=3e-5, epsilon=1e-08),
+                      metrics=['mae'])
+        return model
+
+
+    def get_model(self):
+        if not self.get_classes():
+            return self.get_regression_model()
+        else:
+            return self.get_classifier()